I am optimizing this application for better load times, I noticed the approach in which I took in the backend made it quite slow responding to request

+++ controllers/transaction.js
@@ -1,6 +1,7 @@
+const createError = require("../lib/error");
 const Transaction = require("../schemas/Transaction");
 
-// Create Transaction
+// Create Transaction // userId should now be added to further created transactions on the request
 const createTransaction = async (req, res, next) => {
   const transData = req.body;
 
@@ -19,7 +20,7 @@ const createTransaction = async (req, res, next) => {
             detail: "Transaction successfully created",
           });
     } else {
-      return res.status(200).json({ existingData: existingTransaction.uniqueIdentifier});
+      return res.status(200).json({ existingData: existingTransaction.uniqueIdentifier });
     }
 
   } catch (error) {
@@ -31,27 +32,34 @@ const createTransaction = async (req, res, next) => {
 const updateTransactions = (req, res, next) => {
   const transactionId = req.params.transactionId;
   const newTransaction = req.body;
+  const userId = req.query.userId;
 
-  Transaction.findByIdAndUpdate(
-    transactionId,
-    { $set: newTransaction },
-    { new: true }
-  )
-    .then((transaction) => {
-      res.status(201).json({
-        message: "Success",
-        status: 200,
-        detail: "Transaction successfully updated",
-      });
-    })
-    .catch((err) => {
-      next(err);
-    });
+  Transaction.findById(transactionId).then((transaction) => {
+    if(transaction.userId === userId) {
+      Transaction.findByIdAndUpdate(
+        transactionId,
+        { $set: newTransaction },
+        { new: true }
+      )
+        .then((transaction) => {
+          res.status(201).json({
+            message: "Success",
+            status: 200,
+            detail: "Transaction successfully updated",
+          });
+        })
+        .catch((err) => {
+          next(err);
+        });
+    }
+  }).catch(err => createError(403, 'You are not allowed to delete a transaction you did not create', err));
 };
 
 // Get all transactions
 const getTransactions = (req, res, next) => {
-  Transaction.find()
+  const userId = req.query.userId;
+
+  Transaction.find({ userId })
     .then((transactions) => {
       res.status(200).json(transactions);
     })
@@ -63,10 +71,15 @@ const getTransactions = (req, res, next) => {
 // Get single Transaction
 const getTransaction = (req, res, next) => {
   const transId = req.params.transactionId;
+  const userId = req.query.userId;
 
   Transaction.findById(transId)
     .then((trans) => {
-      res.status(200).json(trans);
+      if(trans.userId === userId) {
+        res.status(200).json(trans);
+      } else {
+        res.status(403).json({ message: 'You are not allowed to request this transaction', status: 403, detail: 'You can only request for transactions you created, our clients confidentiality is our top priority' })
+      }
     })
     .catch((err) => {
       next(err);
@@ -76,20 +89,25 @@ const getTransaction = (req, res, next) => {
 // Delete Transaction
 const deleteTrans = (req, res, next) => {
   const transId = req.params.transactionId;
-
-  Transaction.findByIdAndDelete(transId)
-    .then(() => {
-      res
-        .status(200)
-        .json({
-          message: "Success",
-          status: 200,
-          detail: "Transaction Successfully deleted",
-        });
-    })
-    .catch((err) => {
-      next(err);
-    });
+  const userId = req.query.userId;
+  Transaction.findById(transactionId).then((transaction) => {
+      if(transaction.userId === userId) {
+        Transaction.findByIdAndDelete(transId)
+        .then(() => {
+        res
+          .status(200)
+          .json({
+            message: "Success",
+            status: 200,
+            detail: "Transaction Successfully deleted",
+          });
+      })
+      .catch((err) => {
+        next(err);
+      });
+    }
+  })
+    
 };
 
 module.exports = {
